Add jenkins:sshd to unlimited cgroup.
Jenkins slaves use ssh to communicate with the master node. The
sshd process used to perform this communication should be in the same
unlimited cgroup as jenkins:java. Make it so.

This protects the sshd process from running into limits and gives
the tests a little more headroom as the sshd process won't count
against the tests' limits.

Change-Id: I4a047a2d018f80a64d5f0495f5c57d125fe1119f