Directly create the baremetal test database.
mysql::db assumes it has to create the user given to it each time it is
invoked. This break when you want to give one user access to mulitple
DBs. To work around this directly create the new DB and grant
permissions to it using the types provided by the puppetlabs-mysql
module. This change was made copying what mysql::db does[1].

[1] https://github.com/puppetlabs/puppetlabs-mysql/blob/0.5.0/manifests/db.pp

Change-Id: Ie45c1bd6f276bfbbc776e816cb5957a7fe808882
Reviewed-on: https://review.openstack.org/22393
Reviewed-by: Jeremy Stanley <fungi@yuggoth.org>
Approved: Clark Boylan <clark.boylan@gmail.com>
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Tested-by: Jenkins