Disable logging to /dev/xconsole.
The default rsyslog configuration on Ubuntu attempts to open
/dev/xconsole if it exists, and logs an error if it doesn't (which
it normally won't on a headless server anyway). This change installs
a modified configuration with the xconsole-related lines commented
out. This configuration file is valid at least on Ubuntu Oneiric,
Precise and Quantal, which covers the releases we currently use.

Change-Id: I68594d489ab50ef25d351162b9dcb50ca003c409
Reviewed-on: https://review.openstack.org/22394
Reviewed-by: James E. Blair <corvus@inaugust.com>
Approved: Clark Boylan <clark.boylan@gmail.com>
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Tested-by: Jenkins