Check the number of tests run by testr.
Check that the number of tests run by testr is equal to the number of
tests found by testr. If subunit has a catastrophic failure these two
numbers will not match but testr reports success. Check if the numbers
match and if not fail the build with a useful message.

Change-Id: I849db8912dd0d9974fb83ed08bc93646b0eb3e9f
Reviewed-on: https://review.openstack.org/22456
Reviewed-by: Dan Smith <danms@us.ibm.com>
Reviewed-by: Monty Taylor <mordred@inaugust.com>
Reviewed-by: James E. Blair <corvus@inaugust.com>
Reviewed-by: Jeremy Stanley <fungi@yuggoth.org>
Reviewed-by: Sean Dague <sdague@linux.vnet.ibm.com>
Reviewed-by: Khai Do <zaro0508@gmail.com>
Approved: Clark Boylan <clark.boylan@gmail.com>
Tested-by: Jenkins