Check find's output in Jenkins' translation scripts.
* modules/jenkins/files/slave_scripts/propose_translation_update.sh:
Instead of evaluating find's output in test treat finds output as a
string and check if that string is null to determine if find found any
.po files. This is a more reliable way to check for the presence of .po
files as it does not result in confusing test.

* modules/jenkins/files/slave_scripts/upstream_translation_update.sh:
Instead of evaluating find's output in test treat finds output as a
string and check if that string is null to determine if find found any
.po files. This is a more reliable way to check for the presence of .po
files as it does not result in confusing test.

Change-Id: I53f7cc47e3762e4529eaf72f457852e650413b16
Reviewed-on: https://review.openstack.org/26346
Reviewed-by: James E. Blair <corvus@inaugust.com>
Approved: James E. Blair <corvus@inaugust.com>
Tested-by: Jenkins