Install additional sound prompts
Telephone calls are generally 8 kHz.  With VoIP, higher sample rates
are possible.  G.722 is the most common, which is 16 kHz.  Conferencing
supports these higher sample rates, as well.  Make sure we have the
higher quality sound prompts installed as well in case we can make use
of them.

While we're at it, install the gsm prompts, as well.  That is the low
bandwidth codec currently allowed.

Change-Id: I34921fa6a00720d05113a848bc9f1f94f2200c8b