Update the default extension
This patch updates the default extension.  This is what a caller would
hit if they called sip:openstack.org (as opposed to
sip:6000@openstack.org), or when they are calling in via the PSTN.  It
asks them to enter a conference number and then hit #.  It will give up
after 3 tries and just hang up.

I moved the 'spam' playback that I originally had to only play if you
enter the digits for spam (7726).

Change-Id: Ia82b9b52766ee191e3fd6f20d4e3b7fbb56f1f5b