Workaround tox 1.6 pypi workaround
tox 1.6 tries to be helpful and set your index server for easy_install
as well. (die easy_install die) But, its hack around the problem
conflicts with our hack around the problem. Thus, an incomplete solution
that at least fixes core projects, is to set TOX_INDEX_URL in the
environment so that it will at least set the right thing.

Closes-Bug: 1212751
Change-Id: I0bc29a35fba3b308a7b52778baa575818533ffd5