Fix zuul status hours display.
* modules/openstack_project/files/zuul/status.js: The hours field in the
zuul status was always '00'. This was a result of calculating the
minutes to be displayed rather than the total number of minutes before
calculating the number of hours from the number of minutes. Since
minutes was always less than 60 hours was always 0. Determine the total
number of minutes, calculate the hours, then determine the number of
minutes to display.

Change-Id: I3643282247d21fdd1b56c0baa00380bfdc7ae2e2
Fixes-Bug: #1215659