Use "sort -V" for tags when building docs
The docs script changes its behavior when running on the "latest
release tag" -- it prepares to publish the docs at the primary
location, rather than in a subdirectory.  It used "sort -n", or
a numeric sort, of all tags to determine whether the current tag
was the latest.  This would fail to sort "0.11.0" after "0.9.0".

This changes the script to use a version number sort which
recognizes that difference.

As for comptability with other version number formats: it is
compatible with year.date formats (though the script ignores those
for client libraries).  And versions with letters in them (ie,
pre-release versions) are already filtered out (a pre-release can
never be the latest released tag).

Change-Id: I6bba9bbc97b8c8e635d1b88835129113485baf21