Add saltstack APT repository to automatic upgrades
* modules/openstack_project/manifests/automatic_upgrades.pp: Pass
the additional allowed origins list through to the unattended
upgrades module.

* modules/openstack_project/manifests/slave.pp: Add the
"LP-PPA-saltstack-salt precise" repository origin to the list of
allowed origins for automatic upgrades.

* modules/unattended_upgrades/manifests/init.pp: Add a list of
additional allowed origins for package repository upgrades and
convert the apt-preferences stub for 50unattended-upgrades to a
template so the list can be applied to it.

* modules/unattended_upgrades/templates/50unattended-upgrades.erb:
Renamed from ../../files/50unattended-upgrades and enhanced with a
loop over the origins list to append to the default set of allowed
upgrade origins.

Change-Id: I821d199ae1a00ed78c2a52bf80ddba8a4eaa7d8b