make check queue high priority
we've currently starved check to the point where it's been > 15hrs
since any check entering change has gotten a devstack node. This
starvation has blocked a lot of developers from making progress,
and is causing people to jump the check queue entirely.

It is also preventing any feedback on changes we need to debug the
gate failures.

Put check back on par with gate for now to start making forward
progress.

Change-Id: I593a6735284712c578b4193eeda9bf224a336799