Clean up in requirements job
We don't run this job with a fully clean workspace (git clean in the
requirements repo does not seem to also delete git repositories that
have been cloned as subdirs) So we need to remove and re-clone.

Also, while looking in to this, concurrent: false was missing. I'm
pretty sure we don't want to run more than one of these.

Change-Id: I09d1d07b815c348df5d9ae664b84dba9d5d169f0