create a selection of jobs to not use config drive
This takes a set of our jobs and makes them disable config drive
which then means cloud init in the guests will rely on metadata
server instead. The jobs that were chosen for this

 * nova-network + pg
 * neutron + mysql + tenant_isolation (partial isolation still)
 * neutron + pg w/o tenant_isolation

This scattering should be reasonably representative to feel that
we are exercising config_drive and metadata server well.

Change-Id: Ic2acb568b55a05b84120b181383ea45ea190dcad
Needed-by: Ib203bb0c812939f32b36731464b18566a187c85a