Add welcome_message.py to patchset-created trigger
This patch adds the welcome message script to the patchset-created
trigger in gerrit. In theory it should post a lovely message to
any committer making a patchset for the first time.

However, it has not been tested, and review-dev is not available.
So, this patch uses the --dry-run option of welcome_message.py
so that it will hopefully post useful information to the log, but
won't cause catastrophic failures.

Change-Id: I4f04fcd2bd43f680f0d5032fc613194be8dd2586