Sort projects in layout.yaml
Group by type of project, and within groups, sort more or less
alphabetically (as close as makes sense without splitting up, say
the oslo projects).

In addition to helping avoid merge conflicts with additions, this
also makes maintenance easier as projects which are likely to have
similar configurations are in closer proximity.

Change-Id: Ida5399653b5f1b5281e66adb32563fa0a46a53a5