Chef style testing enablement and minor speed cleanup starting with checks
Broke the chef-cookbook-prep builder into bundler and berkshelf steps because
both are not needed for every test. Renamed builders to reflect the underlying
tools used:

  chef-cookbook-prep  -> chef-bundler-prep
                      -> chef-berkshelf-prep
  chef-cookbook-rspec -> chef-cookbook-chefspec
  chef-cookbook-lint  -> chef-cookbook-foodcritic
  chef-cookbook-style -> chef-cookbook-rubocop

Enabled rubocop/style checks without gating for all cookbooks. Will add gating
once the Rubocop updates are in place and this is shown to work well.

Addresses: blueprint lint-and-unit-testing-for-havana
Change-Id: I894f5aefe650ef799ac1aa10345749d2a86f0074