Make cells job gating
We've had a cells job running for a while now and it  has been working
reliably.  Run it everywhere we are running tempest-dsvm-full (except on
tempest, as that doesn't make any sense since this doesn't use
tempest at all).

The current job uses devstack exercises.  At some point we'll get
this updated to use a subset of tempest that we expect to work with
cells.  That's tracked here:

    https://blueprints.launchpad.net/nova/+spec/cells-tempest

Change-Id: Ib483084bc2371089625caa4279a41c0992adfaef