Make python-swiftclient gate symmetrical
This commit adds the standard set of tempest jobs to the
python-swiftclient set. Previously swiftclient wasn't used by
tempest which is most likely the reason these jobs weren't run on
the queue before. (despite glance using swiftclient by default with
devstack in the gate) So to mitigate the risk of a swiftclient patch
breaking the gate the tempest jobs are run on each commit to swiftclient.

Change-Id: I2adb2e3ce1c69c02ec248cbebbe89818676ac0f4