Increase the timeout for check-tripleo-overcloud
The current timeout, is pretty much the same as the length of time the
test takes, increasing it by 20 minutes ensures we don't timeout jobs
just before they they are about to pass. There is a separate effort
being kicked off to reduce the time it takes this test to run once
successful we can reduce this timeout to something shorter.

Change-Id: Id44c791ca7acf6086299c20074cdd649afa863a9