create an integrated-gate template
start normalizing the jobs by creating an integrated gate template
to definite common jobs we expect all the integrated projects to
run. Start small with this with just

 - tempest-dsvm-full
 - tempest-dsvm-neutron
 - grenade-dsvm

Which should be a non controversial common set of jobs. After this
merges we can take other jobs one at a time here.

Change-Id: I3d96dcd8f96c3ada9d375b28a564dc43d8a88d74