Only attempt to rename tarballs in tarball job.
The branch tarball job for python projects has been failing since we
started writing md5 and sha1 sums to files because it is trying to copy
multiple files to a single destination. Only attempt to rename tar.gz
files so that we have a 1:1 relationship between source tarball and
renamed branch destination tarball.

Change-Id: I744b959717269fc50846efba8639908079e6d0ff