Change names for duplicate neutron jobs
This commit adds a *-2 neutron jobs to be used for running duplicate
jobs for proposed changes to neutron. Originally duplicate jobs with
the same name were used which were added in change:
I33eda4b22505864fb7b609c90703df5fbddf9d71 however there is an issue
with reporting with the jobs names duplicated like that. This works
around the issue by changing the name of the duplicate job.

Change-Id: Iae6980777d6da78f74bf35a4378da7a29acfc336