Disable broken oslo jobs
The gate job for stevedore docs was turned on recently,
but the requirements for building those docs are not allowed
and we are in feature-freeze so we can't add them, yet.
Make the job non-voting so we aren't blocked setting up
library test jobs.

The requirements job for cliff is failing because the master
branch has duplicate requirements and the parser is complaining
before it verifies that an update fixes the problem.

Change-Id: I40b2cf4030df66c851d7d11da4f81f645d110e76