make postgresql job not use config drive
because we had duplicate looking pg jobs in devstack-gate.yaml
when the file injection behavior was changed back in January it
turns out the nova job to test metadata server was lost (as it
was defined on the pg job that we don't run).

Remove the pg smoke job that's not run to avoid confusion, and set
DEVSTACK_GATE_CONFIGDRIVE=0 on the pg job to get us a metadata
testing environment.

Change-Id: Iff0e3d29f95ab701f1572fea6d8599eade40d181