make tempest-dsvm-neutron-heat-slow non voting
tempest-dsvm-neutron-heat-slow is currently running a nearly 50%
failure rate, which is substantially impacting most other projects.
We've identified a few possible causes of the issue however it's
probably best to make the job non voting for now to be able to
work through fixes and not block other projects from merging code.

Change-Id: I36c4bb32d17841304807e6f3b3449deaca0e4601