Switch review.o.o to testmode
Also, cause testmode to disable manage-projects on gerrit servers.

This will cause puppet to stop writing the replication config (but
leave it in place on the production server) and also stop running
manage-projects.  This is so that we can spin up a new gerrit
server without it immediately creating projects or replicating.

Also, remove the gerrit::mysql module so we can replace it with
a trove database.  Again, the existing mysql databases should be
untouched.

Change-Id: I157993297ae8168064ef5b368d42dc93989aef21