use buck from master branch
Buck is used to build Gerrit.  We needed an older version of it to build
Gerrit 2.8 that's why we needed to checkout a special revision of it.
Recently the Gerrit project was updated to dynamically load the correct
version of buck at build time therefore we don't need to checkout and build
that special version of buck anymore.  Now when you build Gerrit the correct
version of buck gets loaded before the Gerrit code gets built.

Change-Id: I4d5dc97571daf0e6d7a4c45cf16125aa9b0b65cb