Add flag for unbound's resolv.conf
diskimage-builder does not run services that it installs, but this
is a problem when we lay down the resolv.conf file in puppet with
the service not running. So, put in a flag which defaults to true
(which is current behavior) that a dib invocation of the puppet
can turn off and then dib can plop the file down at the end of the
process.

Change-Id: I05d89ffacfdaf3563b8cb1460af12f114e1a0340