Fix zero tests ran unittest check.
The unittests attempt to check that more than 0 tests run; however, in
some cases 0 tests will be run and testr last won't list the number of
tests as a number. This causes the sed filter to fail to match resulting
in a rancount of "". If we get a rancount of "" or "0" fail the test.

Change-Id: I1f7ca358c73c071e60ba4b651cf75db18b8fd487