Remove tripleo cross-testing with oslotest
This was added before we realized a separate cross-test job for
every project that uses oslotest was not going to be practical. It
also seems to have issues with the stable branches of the tripleo
projects, so we're just going to remove them for now.  Once we've
come up with a cross-testing strategy for Oslo we will add these
projects back in to the testing matrix.

Change-Id: I5679a135ab4925d9db8471397d052e53a5c93b4b