Reduce tripleo ci jobs
In order to spend some time redeploying/verifying R1 we need to reduce the
demand for tripleo CI so that R2 can handle the load solo. The 2 main
changes here are 1. Move the ubuntu overcloud job into the experimental
queue, from here it can still be run if needed. 2. Move the
ironic-undercloud job into the experimental queue, it still isn't
passing and can be run from here if needed. We may need to remove more
but this will be a good start.

Change-Id: I78d8d3743598cc725632e4e8fb055e6a73886d66