Determine wheel names for PyPI uploading
* .../files/slave_scripts/pypi-extract-name.py: Add --tarball and
--wheel switches to return the proper corresponding name versions,
since they differ in some cases in the presence of non-alphanumeric
characters. Make --tarball the default for backward-compatibility.
This adds a dependency on the wheel module.

* .../files/slave_scripts/pypi-extract-universal.py: Use the correct
py2 indicator for non-universal wheel types rather than the old py27
fallback.

* .../files/slave_scripts/pypi-tarball-upload.sh: Request the
tarball version of the package name prefix.

* .../files/slave_scripts/pypi-wheel-upload.sh: Request the wheel
version of the package name prefix, and also correct the
version/type separator character.

* .../manifests/pypi_slave.pp: Always install the latest versions of
the wheel and twine packages.

Change-Id: I2ae724f1f4d0a88b602e3581a0b06b999fb4833c