Remove dep on wget package for apt::key
Currently, the apt::key resource has a dep
set for Package['wget']. This dependency
isn't actually a requirement for apt::key for
the version of apt being used by config. The
provider uses native ruby libraries for ftp/http
(open-uri) and thus does not require the wget
package.

Also, having the jenkins modules depend on the wget
package is problematic b/c the resource is defined
in the openstack_project module. As we decouple
out the modules, the non-openstack specific module
(like jenkins) should not have dependencies on the
openstack specific/composition modules (like
openstack_project).

Change-Id: Ia70fa4f4d9c7b2d5f6b3dd476b5b83498049e524