Remove freshness check from tripleo
It doesn't work as expected because it depends on the timing of
the vote from the check queue.  It's unlikely to ever trigger
on its own unless the check queue is 48 hours backlogged (120-72).
Except now it _is_ triggering in a loop because we stopped doing
freshness checks for -1 votes, so we can have an old jenkins vote
that never gets updated.

Change-Id: Ib89ed3c8f76ecded9f04bbcd784c3e91a835cc54