Added NodeJS and NPM to javascript build
It turns out that trusty has a recent-enough version of node
and NPM to allow us to remove the nodeenv dependency in
our javascript builds. This is the first step in doing this, to
get the core javascript build tools onto our jenkins nodes.

Note that a symlink for node is also created, because debian
has a different module called "node" for which there's a
namespace conflict. Since it's isolated to the npm-install
macro and we're unlikley to ever do Amateur Radio things
on infra, I don't expect this to be a problem.

Change-Id: I5349ad2013185fa20e784e3aa3691c7d41c43636