Move gate configuration to nova-docker.
We're at the stage now where nova-docker tempest CI is getting close
to passing locally, so we would like to enable full tempest runs for
each code commit.  We will do this by setting the environment
variables in nova-dockers gate hook override.  Also clean up cp of
script.

The post_test_hook was removed as it's no longer used. It previously
ran a trivial smoke test that we are now replacing with a full tempest
run.

Change-Id: I1e1983f3f03e72ba7eba80fb81426745d08e54e1