Remove duplicate unused job definitions
Remove duplicate references to 'gate-{name}-docs' where the project uses
the job-group 'python-jobs' as this already includes the docs job.

Remove template-job definition for 'gate-{name}-build' as it has been
replaced by a job-group which supersedes the template-job definition due
to the parse order of jobs, job groups and job templates.

Change-Id: I096087e2e6663f919fb3779a7b628abde2ebe0ab