Publish additional specs repositories
Remove -python27 jobs from specs-jobs template so that we can use
the template everywhere we publish specs. Adjust all current
users of that job

Use the template for all jobs that already gate the docs.
This is done only for "official" OpenStack repositories but
not for stackforge ones.

Also, add python27 job for glance-specs and use the template. 'tox -e
python27' is setup properly and passes.

For cinder-specs, use the specs-jobs template in both projects.yaml and
layout.yaml.

Change-Id: Iba0badd60a70e467e076c8c829c1e8d0dd6b1591