Sync files from openstack-manuals with other repos
Rename propose_requirements_update.sh to propose_update.sh and
adjust it so that it handles syncing of files for openstack-manual
as well.

The files to be copied are right now the glossary files and they are
copied to operations-guide and security-doc. The copying of the glossary
allows to use openstack-manuals as master reference and also sharing of
translations.

Adjust jjb config of propose_requirements job.
Add new proposal job for openstack-manuals and run it as post job.

implements bp common-glossary-setup

Change-Id: I3f2d6238c17f45dd149ec282270227a14c7dd1b1