add experimental job for nova with f20 virt-preview
This patch adds an experimental job to nova that uses the Fedora 20
node type that has the virt-preview repository enabled.  This will run
the usual full tempest suite, but with the latest versions of libvirt
and qemu installed.

This revision of the patch unfortunately duplicates the jenkins job
config.  It seems like that shouldn't be necessary, but I haven't
figured out the right way to massage that into place just yet.

Change-Id: I9d0d7feb68bd16997a5d9fc4a639aec4473ee2b4