Add neutron full job to integrated gate
This change will add the neutron full job to the integrated gate,
leaving it as non-voting.

As time is ripe to switch this job to voting, it makes sense to add
it, in its current form, to the integrated gate so that it's executed
also with projects different from neutron, devstack and tempest.

Change-Id: I7f2c82c994d7fc93733deb1f82c6910335589b85