Promote Nova check for docker to silent
It has been the policy of the Nova project that virt-drivers
must prove a positive history of functional testing prior to
being introduced (or re-introduced) to the project.

This change promotes the experimental nova gate to a silent job
which will run a non-voting check against the nova-docker driver
to provide valuable statistics and history necessary for eventual
re-introduction.

Furthermore, this will help improve the quality of the nova-docker
driver and aid maintenance as we approach the Kilo window.

Discussion on this change occurred on the mailing list:
 http://lists.openstack.org/pipermail/openstack-dev/2014-August/043183.html

Change-Id: I74e0bda10e85fbe430238f9b53d53b3671309a6a