Pin puppet lint to 0.3.2.
Puppet lint released a 1.0 version that we are not prepared for as
existing checks are not more strict. Pin to the old version that we work
against allowing us to continuing to use puppet lint in a way that works
for us. We can consider upgrading at some point in the future.

Change-Id: I310c9d4e09f706f24e67f39ec6fbb1fb006148dc