Update RH1 to use net-label instead of net-id
Nodepool now supports net-labels so lets start by using
them on the RH1 TripleO cloud. The existing nodepool
implementation rely's on novaclient (i.e. a nova extension)
which is deployed via TripleO and should work fine.
Future work may include the use of neutronclient which
may enable us to use this approach for other (public)
clouds.

The motivation here is to make nodepool.yaml a bit more
readable with regards to the net-id's.

Change-Id: Icc1c44803928a13790805837dedb3c0fb15e9cbe