Remove havana compat jobs from tempest experimental
This commit removes the havana jobs from the tempest experimental
pipeline. Those jobs were only added to test how much work was needed
to make tempest master work against stable/havana of the projects. It
was determined that it would be too much effort to backport the fixes
needed to make this work, and with the pending eol of havana we should
just remove the jobs.

Change-Id: I508c72c40b43c5cfc943dd2521c07caa08c5ad6f