Allow puppet to be automatically upgraded
Since the puppet 3 release puppet versions are semver and we should be
able to rely on minor version releases not breaking the world. To get
security updates and the like allow puppet to be automatically upgraded.
Do this by setting an additional origin for the puppetlabs apt repo in
our unattended upgrades template.

Change-Id: I6224ef68aeb891431661f3afc553c4c3433e583a