Run bandersnatch under file locking and a timeout
* modules/openstack_project/manifests/pypi_mirror.pp: Bandersnatch
occasionally seems to livelock in a select poll loop at teardown.

https://bitbucket.org/pypa/bandersnatch/issue/52

As such this blocks subsequent runs of our wrapper script, filling
the log with useless tracebacks every 5 minutes while leaving the
mirror perpetually stale until someone comes along and notices. Work
around this unfortunate situation by setting a half hour timeout on
the process, and add a file lock to the invocation so that it will
not try to run more instances of the wrapper script until the
current one terminates.

Change-Id: I055b623801af3063d22b081eb9c9ef3634e5e5b5