Add devstack-trusty-dib image.
Add a new devstack trusty image to nodepool that is built via dib on the
nodepool server. This is our first test of the new dib features added to
nodepool. Once we have these images in glance and are able to boot nodes
off of them we can add new tests that check their behavior.

Change-Id: I179df9117be538c3ab44cce3a3cd579e03e21e90