Set the nodepool user and key on dib images
We need to set the nodepool user and key on dib images so that dib can
run the pre use check script on the nodes (without these values nodepool
uses invalid defaults). Logically this appears to be part of the image
building process but the post slave boot check scripts need it too.

Change-Id: I51c7da1d58c2550903ad1671dcfb477b06424cfd