Temporarily disable access to openstackid.org
The production OpenStackID service is able to read sensitive data,
so before booting the production server we're temporarily removing
global Web access to the system and will manually whitelist specific
source addresses for testing until we're comfortable it's safe to
open up to the World. At that point, this commit will be reverted.

Change-Id: I7a89fa1ffccaf777c29fb22e06c6c14929739657