Add centos7 to nodepool diskimage config
I accidentially dropped this section of the config file when rebasing
between patchset 3 & 4 of change
Ic1848b98625603bf37c0e3dfaf1d0ef6f2c6ac74

Note this is slightly different in that the release is not specified,
so we take the default centos7 image from d-i-b.  This is no different
(at the moment they are the same) but means if upstream d-i-b changes
their release, nodepool follows "silently".  We will monitor to see if
this is a feature or a bug.

Change-Id: I0c7738f59d014238934b6448ae7bb591eee1bcbe