Set devstack-centos7-dib min ready to 0
We cannot boot the centos7-dib image we have today because it is based
on a broken base image and we cannot build a new working image until we
can run dib on a node with tar that supports --xattrs. In the mean time
set the min ready to 0 so that we stop booting VMs that are doomed to
fail at boot.

Change-Id: Ieaea0f04c02d09f3d0f6d649180607d2a0fd87e6