Puppet the puppetmaster apache vhost
We need to slightly tweak the puppetmaster-passenger package's apache
vhost file slightly for our environment. First we need to set a max
requiests limit for passenger processes so that they are cycled out in
order to avoid a memory leak. Second we enforce TLS and no SSL to
prevent POODLE.

Change-Id: I309d62866a7706be1ae3bedbf45ab9ffb8e04e50