Rotate out nodepool logs every 8-hours
Individual nodepool .log files have now grown to being > 250mb.  This
is a bit much to download and go through when you're trying to debug
an issue with a particular image.

8-hour rotations seems a bit more useful than 24-hour ... and perhaps
even that is a bit long.  I don't think we need to extended the number
of rotations, I'm not sure really old logs are that helpful.

Change-Id: Ia788db809afe470bb9241f17411e4fa97d09ea11