only show most recent results per pipeline
Change hideci to only show the most recent results of a pipeline. For
entries where the pipeline is not parsable we default to assuming the
results were in the 'check' pipeline.

3rd Party CI systems are assumed to be 'check' pipeline for the
purposes of this patch.

Order the display of these results for parsed pipeline entries first,
then non parsed entries after. That has the impact of making all the
Jenkins results be the top of the page and 3rd Party come later.

Also provide a (### rechecks) comment if we find more than 1 result at
the current patch set number. This helps reviewers understand that
this patch might be unstable as it's been rechecked to get to the
current state.

Change-Id: I153f73000f9392af6c8bbe850716645b3ba836a0